Import live.2003.01.17.tar.gz
2003.01.16
Improved support for receiving MPEG-4 LATM audio.  Note that currently,
this works only for streams that do not have a StreamMuxConfig present
in the stream.

2003.01.17
Further improvements to the MPEG-4 audio support.  We can now read
(from the "MediaSubsession") the StreamMuxConfig "config" stream that
was present in the SDP description.  Also added a routine that parses
such a string, producing binary "AudioSpecificConfig" data.
(See "MPEG4LATMAudioRTPSource.cpp" for details.)

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>