Import live.2002.12.10.tar.gz
2002.11.22
Updated "QuickTimeFileSink" once again - this time to add more statistics
to the 'udta'/'hinf' atom for each generated hint track.

2002.11.25
Another change to "QuickTimeFileSink".  Now, when synchronzed tracks
are requested, the durations of video frame samples are adjusted so
that they correspond to actual RTP presentation times.  (This is not
done for audio samples, however, because having audio samples vary
in duration might break audio codecs.)

2002.11.30
Renamed "TaskScheduler::blockMyself()" to "doEventLoop()", to better
describe what this member function actually does.

2002.12.10
Yet another change to "QuickTimeFileSink", this time to make sure that
the duration of video tracks corresponds exactly to the sum of video
sample durations listed in the 'stts' atom.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>