Import live.2002.12.21.tar.gz
2002.12.20
- Another change to "QuickTimeFileSink", again to change the way that track
durations are computed for ".mov" files.  Now, each track's duration is
set to be the maximum of (i) the sum of the sample durations listed
in the 'stts' atom, and (ii) the the sum of the durations listed in the
track's 'edit list' (if any).
- Updated the MP3 file parsing code to check for a 'Xing' VBR header in
the first frame.  ("liveCaster" can use this information to print more
sensible information about each file that it streams.)

2002.12.21
Fixed a bug in the previous update to "QuickTimeFileSink"

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>