Import live.2003.02.17.tar.gz
More work on the experimental '-R <rtsp-url>' option to "playSIP"
(and "openRTSP") that allows it to inject the incoming stream into
a separate RTSP server.  This code is now working, although
currently only for a single PCMU or GSM audio stream.  Also, RTCP
packets are not yet relayed between the source and destination.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>