Import live.2003.02.06a.tar.gz
2003.02.03b
Added an #include to "BasicUsageEnvironment/Lock.cpp" to fix a problem
that someone encountered with NULL not being defined.

2003.02.04
Removed "-DUSE_OUR_BZERO=1" from "config.linux", because it no longer
seems to be necessary, and it was breaking compilation for some people.

2003.02.05
- Cleaned up "RTSPClient", and added a method for doing an aggregate
"PLAY" operation (in addition to the existing method that does a
"PLAY" on an individual subsession).
- Modified the "openRTSP" test program so that it does a single
aggregate "PLAY", rather than a series of non-aggregate "PLAY"s, one for
each subsession.  (RealNetworks' server doesn't support the latter.)
- Added a new routine for parsing "AudioSpecificConfig" strings that
can appear in SDP descriptions.
- Added Morgan Multimedia's implementation of "JPEGVideoRTPSource"
to the "liveMedia" library.  (I haven't yet tested this.)

2003.02.06
Fixed some code that was 64-bit-unsafe.  (Thanks to Philipp Thomas for
noticing this.)

2003.02.06a
- Fixed a bug in the SDP "config" attribute parsing.  (This affected
the use of "MPEG4LATMAudioRTPSource".
- Cleaned up the "createNew()" routines in each of the *RTPSource classes.
- Added two new RTPSource classes - "MPEG4ESVideoRTPSource" and
"MPEG4GenericRTPSource", for MPEG-4.  However, these have not been
fully implemented yet.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>