Import live.2003.03.22.tar.gz
2003.03.07
- Added an optional "MIME subtype" parameter to "SIPClient", to allow
SIP "INVITE"s to use a dynamic RTP payload format.  Updated "playSIP"
accordingly, to add an optional "-D <mime-subtype-name>" argument.
- Modified "FileSink" to close the source (and stop playing) if it gets
an EOF when writing to the output file.  (Thanks to Ruud Schramp.)

2003.03.09
- Made "MediaSession"s parsing of SDP descriptions more robust, to allow for
blank lines in the SDP description.  ("Be liberal in what you accept...")

2003.03.11
Updated SIPClient to allow it to use an arbitrary source port number.
This makes it possible to use it on the same host on which a SIP server
is already running.

2003.03.14
- Removed calls to "strdup()", because that's a C-library function whose
result should not be deleted using "delete" (or "delete[]").  Instead,
we now provide our own C++ equvalent, called "strDup()".
- Also, changed several "delete"s to "delete[]".
(Thanks to Bill Kain for noting the need for this fix.)

2003.03.22
Added support for receiving H.261/RTP streams

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>