Import live.2003.04.04.tar.gz
2003.04.01
Fixed a bug in "openRTSP" that could cause it to crash when exiting
(if the "-t" option was used).

2003.04.02
- Made another change to the order that things are closed when exiting
"openRTSP", to help avoid a potential race condition that can occur
when using the "-t" option.  (Thanks to Romulus Grigoras for reporting
this.)
- Updated "MPEGDemux" to properly handle the case where someone tries
to read from it after its input source has already closed.
- Several improvements to "VOBStreamer".
- Added "ByteStreamMultiFileSource" to the "liveMedia" library.
This generalizes "ByteStreamFileSource" to allow more than one input
file to be read, in sequence.

2003.04.04
- More improvements to "vobStreamer" (which has also been renamed
from "VOBStreamer".

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>