Import live.2003.07.14.tar.gz
2003.07.13
- Added new filter classes for converting between 16-bit PCM and
8-bit u-law audio, and between host and network order for 16-bit values.
(See "liveMedia/include/uLawAudioFilter.hh" for details.)
- Added a new class "WAVAudioFileSource" for reading/processing
WAV audio files.
- Added a new test program "testWAVAudioStreamer" that reads a WAV
audio file - "test.wav" - and streams it via multicast (using a built-in
RTSP server).
- Updated "MediaSession" to recognize "L8" and "L16" RTP audio streams
(to be handled by "SimpleRTPSource").

2003.07.14
Updated the "WAVAudioFileSource" implementation (and the
"testWAVAudioStreamer" test program) to better report any reason for
a WAV file to be invalid.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>