Import live.2003.07.07.tar.gz
2003.05.12
- Restructured "BasicTaskScheduler" as two hierarchical classes
  "BasicTaskScheduler0" (an abstract base class), and
  "BasicTaskScheduler".
  This makes it easier to subclass (e.g., to reimplement "SingleStep").
  A subclass reimplementation would use "BasicTaskScheduler0".
- Removed "our_bcopy()", and use "memmove()" instead, because everyone
  seems to implement that.
- Fixed a bug in "RTPSource" that could cause packet loss stats to be
  reported incorrectly if the first RTP sequence number received was 0.

2003.05.15
- Moved the "select()" call in "readSocket()" (GroupsockHelper.cpp) into
  a separate function, to allow it to be easily reimplemented
  if desired.
- More minor changes to "RTPSource" stats handling, and the "-Q" option
  to "openRTSP"/"playSIP".

2003.05.16
- Removed all calls to "fprintf(stderr, ...)" and "cerr << ..." from
the library code.  Instead, we now use "operator<<" virtual functions
that are defined on the "UsageEnvironment".  "BasicUsageEnvironment"
defines these by outputting to "cerr", but other subclasses of
"UsageEnvironment" can define them to do console output whichever
way they wish.

2003.05.19
Fixed a bug in "SIPClient" that could cause a crash.  (Thanks to
Reini Urban for finding this.)

2003.05.21
- Because not all platforms define 'errno' the same way, 'errno' is now
implemented as a pure virtual function "getErrno()" in
"UsageEnvironment".  "BasicUsageEnvironment" defines this function to
just return "errno", but other subclasses could implement it in other
ways.
- Added support for setting and reading information in "RTP-Info:"
headers in RTSP "PLAY" response. (Thanks to Romulus Grigoras for
contributing this.)

2003.05.22
Fixed a bug in "SIPClient" (a field was not getting initialized properly)

2003.05.22a
Fixed a couple more minor bugs that were causing "valgrind" to complain.

2003.05.22b
Fixed a bug in "SIPClient" (thanks to Reini Urban for helping track this
down).

2003.05.23
Some minor changes to reduce the number of bogus warning messages
displayed when compiling using Visual C++.

2003.05.28
Fixed some typos in comments.

2003.05.29
Fixed a bug in "RTSPServer" that would cause it to get into an infinite
loop if a client's connection was terminated unexpectedly.

2003.05.31
Fixed a bug in "SimpleRTPSink" that would cause it to not generate correct
RTP timestamps on outgoing packets.

2003.06.07
- Fixed a minor bug in "MPEGVideoRTPSink": The RTP "M" bit is now set
only on the last fragment, if a frame is fragmented.  (Thanks to
Tym Altman for pointing this out.)
- Improved support for AudioRTPSinks that have more than one audio channel.

2003.06.13
- Fixed a bug in "ServerMediaSubsession" that could cause a memory leak.
(Thanks to Frederik Bonte for finding this.)

2003.06.17
Added support for MPEG-4 generic RTP sinks.  (Currently, only the
"AAC-hbr" audio mode is supported, and only one AAC frame is packed
into each outgoing RTP packet.)

2003.06.26
- Updated the "RTSPServer" implementation to respond to "PAUSE"
requests.  (It doesnt actually pause the stream; just returns "OK".
This makes QuickTime Player happier.)
- Added a "-o" option to "openRTSP" (and modified "RTSPClient"
accordingly).  This option causes openRTSP to send an "OPTIONS" command
to the server, and prints out the response.

2003.06.27
Added a "AudioDeviceSource" class to the "liveMedia" library.
This class is a generic audio input device (such as a microphone
or a sound card) - to be subclassed for specific OS platforms.

2003.06.28
Added an optional parameter (default, True) to "SimpleRTPSource"
that says whether to use the RTP "M" (marker) bit to indicate the
last (or only) fragment of a video frame.  (Suggested by
Cezar Plesca.)

2003.07.01
Added a new subdirectory "WindowsAudioInputDevice", which is a
Windows-specific implementation of the "AudioInputDevice" class.
This can be used by Windows applications to read PCM audio from
a sound card.

2003.07.01a
Updated the "WindowsAudioInputDevice" subdirectory to
- Change the name of the library from "WindowsAudioInputDevice.lib" to
"libWindowsAudioInputDevice.lib".
- Added a console test program - "showAudioInputPorts" - that lists
the audio input ports that are currently available.

2003.07.07
- Fixed a bug that would cause "vobStreamer" to crash if (incorrectly)
run with no arguments.
- Updated the "test*Streamer" test programs to print out the "rtsp://"
URL of the built-in RTSP server (if it's enabled).

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>